wip: impl Default trait for PPU

+++ crates/ppu/src/config.rs
@@ -1,5 +1,5 @@
 pub const DOTS_PER_SCANLINE: u16 = 456;
 pub const SCANLINES_PER_FRAME: u8 = 154;
-pub const RESOLUTION_Y: u8 = 144;
-pub const RESOLUTION_X: u8 = 160;
+pub const RESOLUTION_Y: usize = 144;
+pub const RESOLUTION_X: usize = 160;
 pub const COLOR_PALETTES: [u32; 4] = [0xFFFFFFFF, 0xFFAAAAAA, 0xFF555555, 0xFF000000];
